Drop all possible scratch tables

Drop all possible scratch tables

dropAllScratchTables(
  connectionDetails,
  scratchDatabaseSchema,
  tempAchillesPrefix = "tmpach",
  numThreads = 1,
  tableTypes = c("achilles"),
  outputFolder,
  verboseMode = TRUE,
  defaultAnalysesOnly = TRUE
)

Arguments

connectionDetails

An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

scratchDatabaseSchema

string name of database schema that Achilles scratch tables were written to.

tempAchillesPrefix

The prefix to use for the "temporary" (but actually permanent) Achilles analyses tables. Default is "tmpach"

numThreads

The number of threads to use to run this function. Default is 1 thread.

tableTypes

The types of Achilles scratch tables to drop: achilles

outputFolder

Path to store logs and SQL files

verboseMode

Boolean to determine if the console will show all execution steps. Default = TRUE

defaultAnalysesOnly

Boolean to determine if only default analyses should be run. Including non-default analyses is substantially more resource intensive. Default = TRUE

Details

Drop all possible Achilles scratch tables