exportAO {Achilles}	R Documentation

exportAO

Description

exportAO Exports Achilles statistics - ares option

Usage

exportAO(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  vocabDatabaseSchema,
  outputPath,
  reports = c()
)

Arguments

connectionDetails
An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)

cdmDatabaseSchema
Name of the database schema that contains the OMOP CDM.

resultsDatabaseSchema
Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema

vocabDatabaseSchema
string name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

outputPath
A folder location to save the JSON files. Default is current working folder

reports
vector of reports to run, c() defaults to all reports

See showReportTypes for a list of all report types

Details

Creates export files

Value

none

[Package Achilles version 1.7 Index]