exportPersonToJson {Achilles}	R Documentation

exportPersonToJson

Description

exportPersonToJson Exports Achilles Person report into a JSON form for reports.

Usage

exportPersonToJson(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  outputPath = getwd(),
  vocabDatabaseSchema = cdmDatabaseSchema
)

Arguments

connectionDetails
An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)

cdmDatabaseSchema
Name of the database schema that contains the vocabulary files

resultsDatabaseSchema
of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema

outputPath
folder location to save the JSON files. Default is current working folder

vocabDatabaseSchema
name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

Details

Creates individual files for Person report found in Achilles.Web

Value

none

Examples

## Not run: 
connectionDetails <- DatabaseConnector::createConnectionDetails(dbms = "sql server",
                                                                server = "yourserver")
exportPersonToJson(connectionDetails,
                   cdmDatabaseSchema = "cdm4_sim",
                   outputPath = "your/output/path")

## End(Not run)

[Package Achilles version 1.7 Index]