exportResultsToCSV {Achilles}	R Documentation

exportResultsToCSV

Description

exportResultsToCSV exports all results to a CSV file

Usage

exportResultsToCSV(
  connectionDetails,
  resultsDatabaseSchema,
  analysisIds = c(),
  minCellCount = 5,
  exportFolder
)

Arguments

connectionDetails
An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

resultsDatabaseSchema
Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.

analysisIds
(OPTIONAL) A vector containing the set of Achilles analysisIds for which results will be generated. If not specified, all analyses will be executed. Use getAnalysisDetails to get a list of all Achilles analyses and their Ids.

minCellCount
To avoid patient identification, cells with small counts (<= minCellCount) are deleted. Set to 0 for complete summary without small cell count restrictions.

exportFolder
Path to store results

Details

exportResultsToCSV writes a CSV file with all results to the export folder.

[Package Achilles version 1.7 Index]