exportVisitDetailToJson

exportVisitDetailToJson Exports Achilles VISIT_DETAIL report into a JSON form for reports.

exportVisitDetailToJson(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  outputPath = getwd(),
  vocabDatabaseSchema = cdmDatabaseSchema
)

Arguments

connectionDetails

An R object of type ConnectionDetail (details for the function that contains server info, database type, optionally username/password, port)

cdmDatabaseSchema

Name of the database schema that contains the vocabulary files

resultsDatabaseSchema

Name of the database schema that contains the Achilles analysis files. Default is cdmDatabaseSchema

outputPath

folder location to save the JSON files. Default is current working folder

vocabDatabaseSchema

name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

Value

none

Details

Creates individual files for VISIT_DETAIL report found in Achilles.Web

Examples

if (FALSE) {
connectionDetails <- DatabaseConnector::createConnectionDetails(dbms = "sql server",
                                                                server = "yourserver")
exportVisitDetailToJson(connectionDetails,
                        cdmDatabaseSchema = "cdm4_sim",
                        outputPath = "your/output/path")
}