getSeasonalityScore {Achilles}	R Documentation

Get the seasonality score for a given monthly time series

Description

The seasonality score of a monthly time series is computed as its departure from a uniform distribution.

Usage

getSeasonalityScore(tsData)

Arguments

tsData
A time series object.

Details

The degree of seasonality of a monthly time series is based on its departure from a uniform distribution. If the number of cases for a given concept is uniformly distributed across all time periods (in this case, all months), then its monthly proportion would be approximately constant. In this case, the time series would be considered "strictly non-seasonal" and its "seasonality score" would be zero. Similarly, if all cases recur at a single point in time (that is, in a single month), such a time series would be considered "strictly seasonal" and its seasonality score would be 1. All other time series would have a seasonality score between 0 and 1. Currently, only monthly time series are supported.

Value

A numeric value between 0 and 1 (inclusive) representing the seasonality of a time series.

[Package Achilles version 1.7.2 Index]