getTemporalData {Achilles}	R Documentation

getTemporalData

Description

getTemporalData Retrieve specific monthly analyses data to support temporal characterization.

Usage

getTemporalData(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  analysisIds = NULL,
  conceptId = NULL
)

Arguments

connectionDetails
An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

cdmDatabaseSchema
Fully qualified name of database schema that contains OMOP CDM schema. On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_instance.dbo'.

resultsDatabaseSchema
Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.

analysisIds
(OPTIONAL) A vector containing the set of Achilles analysisIds for which results will be returned. The following are supported: 202,402,602,702,802,1802,2102. If not specified, data for all analysis will be returned. Ignored if conceptId is given.

conceptId
(OPTIONAL) A SNOMED concept_id from the CONCEPT table for which a monthly Achilles analysis exists. If not specified, all concepts for a given analysis will be returned.

Details

getTemporalData Assumes achilles has been run.

Currently supported
Achilles monthly analyses are: 202 - Visit Occurrence 402 - Condition occurrence 602 - Procedure
Occurrence 702 - Drug Exposure 802 - Observation 1802 - Measurement 2102 - Device

Value

A data frame of query results from DatabaseConnector

Examples

## Not run: 
pneumonia <- 255848
monthlyResults <- getTemporalData(connectionDetails = connectionDetails,
                                  cdmDatabaseSchema = "cdm",

  resultsDatabaseSchema = "results", conceptId = pneumonia)

## End(Not run)

[Package Achilles version 1.7 Index]