isStationary {Achilles}	R Documentation

Determine whether or not a time series is stationary in the mean

Description

Uses the Augmented Dickey-Fuller test to determine when the time series has a unit root.

Usage

isStationary(tsData)

Arguments

tsData
A time series object.

Details

A time series must have a minimum of three complete years of data. For details on the implementation of the Augmented Dickey-Fuller test, see the tseries package on cran.

Value

A boolean indicating whether or not the given time series is stationary.

[Package Achilles version 1.7.2 Index]