listMissingAnalyses {Achilles}	R Documentation

listMissingAnalyses

Description

listMissingAnalyses Find and return analyses that exist in getAnalysisDetails, but not in achilles_results or achilles_results_dist

Usage

listMissingAnalyses(connectionDetails, resultsDatabaseSchema)

Arguments

connectionDetails
An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

resultsDatabaseSchema
Fully qualified name of database schema that contains achilles_results and achilles_results_dist tables.

Value

A dataframe which is a subset of getAnalysisDetails

Examples

## Not run: 
Achilles::listMissingAnalyses(connectionDetails = connectionDetails,
                              resultsDatabaseSchema = "results")

## End(Not run)

[Package Achilles version 1.7 Index]