sumAcrossYears {Achilles}	R Documentation

For a monhtly time series, compute sum and proportion by month across all years

Description

For a monhtly time series, compute sum and proportion by month across all years

Usage

sumAcrossYears(tsData)

Arguments

tsData
A time series object

Value

A data frame reporting the monthly sum across all years and the proportion this sum contributes to the total.

[Package Achilles version 1.7.2 Index]