tsCompleteYears {Achilles}	R Documentation

Trim a monthly time series object to so that partial years are removed

Description

Trim a monthly time series object to so that partial years are removed

Usage

tsCompleteYears(tsData)

Arguments

tsData
A time series object

Details

This function is only supported for monthly time series

Value

A time series with partial years removed.

[Package Achilles version 1.7.2 Index]