createIndices {Achilles}	R Documentation

Create indicies

Description

Create indicies

Usage

createIndices(
  connectionDetails,
  resultsDatabaseSchema,
  outputFolder,
  sqlOnly = FALSE,
  verboseMode = TRUE,
  achillesTables = c("achilles_results", "achilles_results_dist")
)

Arguments

connectionDetails
An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

resultsDatabaseSchema
Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.

outputFolder
Path to store logs and SQL files

sqlOnly
TRUE = just generate SQL files, don't actually run, FALSE = run Achilles

verboseMode
Boolean to determine if the console will show all execution steps. Default = TRUE

achillesTables
Which achilles tables should be indexed? Default is both achilles_results and achilles_results_dist.

Details

Post-processing, create indices to help performance. Cannot be used with Redshift.

[Package Achilles version 1.7 Index]