optimizeAtlasCache {Achilles}	R Documentation

Optimize atlas cache

Description

Optimize atlas cache

Usage

optimizeAtlasCache(
  connectionDetails,
  resultsDatabaseSchema,
  vocabDatabaseSchema = resultsDatabaseSchema,
  outputFolder = "output",
  sqlOnly = FALSE,
  verboseMode = TRUE,
  tempAchillesPrefix = "tmpach"
)

Arguments

connectionDetails
An R object of type connectionDetails created using the function createConnectionDetails in the DatabaseConnector package.

resultsDatabaseSchema
Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.

vocabDatabaseSchema
String name of database schema that contains OMOP Vocabulary. Default is cdmDatabaseSchema. On SQL Server, this should specifiy both the database and the schema, so for example 'results.dbo'.

outputFolder
Path to store logs and SQL files

sqlOnly
TRUE = just generate SQL files, don't actually run, FALSE = run Achilles

verboseMode
Boolean to determine if the console will show all execution steps. Default = TRUE

tempAchillesPrefix
The prefix to use for the "temporary" (but actually permanent) Achilles analyses tables. Default is "tmpach"

Details

Post-processing, optimize data for atlas cache in separate table to help performance.

[Package Achilles version 1.7 Index]